refac: rename tools pkg to utils
