+++ pkg/utils/ctx.go
@@ -1,0 +1,3 @@
+package utils
+
+type CTXKey string
