+++ pkg/utils/google_auth.go
@@ -1,29 +1,23 @@
-package sync
+package utils
 
 import (
-	"bytes"
 	"context"
 	"encoding/base64"
-	"encoding/json"
 	"strings"
 	"time"
 
 	"golang.org/x/oauth2/google"
 )
 
-type ctxKey struct {
-	string
-}
-
 const Oauth2User = "_oauth2_"
 
-// isPermanentServiceAccountToken returns true if user is a Google permanent service account token
-func isPermanentServiceAccountToken(registry string, username string) bool {
+// IsGCRPermanentServiceAccountToken returns true if user is a Google permanent service account token
+func IsGCRPermanentServiceAccountToken(registry string, username string) bool {
 	return strings.Contains(registry, ".gcr.io") && strings.Compare(username, Oauth2User) == 0
 }
 
-// gcpTokenFromCreds creates oauth2 token from permanent service account token
-func gcpTokenFromCreds(creds string) (string, time.Time, error) {
+// GCPTokenFromCreds creates oauth2 token from permanent service account token
+func GCPTokenFromCreds(creds string) (string, time.Time, error) {
 	b, err := base64.StdEncoding.DecodeString(creds)
 	if err != nil {
 		return "", time.Time{}, err
@@ -40,14 +34,4 @@ func gcpTokenFromCreds(creds string) (string, time.Time, error) {
 	}
 
 	return token.AccessToken, token.Expiry, nil
-}
-
-func manifestEqual(m1, m2 []byte) bool {
-	var a bytes.Buffer
-	_ = json.Compact(&a, m1)
-
-	var b bytes.Buffer
-	_ = json.Compact(&b, m2)
-
-	return a.String() == b.String()
 }
