+++ pkg/utils/url_test.go
@@ -1,4 +1,4 @@
-package tools
+package utils
 
 import (
 	"testing"
